# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server for SUT SE 2563",
        "title": "SUT SA Example API Playlist Vidoe",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/institutions": {
            "get": {
                "description": "list institutionentities",
                "produces": [
                    "application/json"
                ],
                "summary": "List institution entities",
                "operationId": "list-institution",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Limit",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Offset",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/ent.Institution"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            }
        },
        "/institutions/{id}": {
            "get": {
                "description": "get institution by ID",
                "produces": [
                    "application/json"
                ],
                "summary": "Get a institution entity by ID",
                "operationId": "get-institution",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Institution ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ent.Institution"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "ent.Activity": {
            "type": "object",
            "properties": {
                "ACTIVITYNAME": {
                    "description": "ACTIVITYNAME holds the value of the \"ACTIVITYNAME\" field.",
                    "type": "string"
                },
                "ADDED": {
                    "description": "ADDED holds the value of the \"ADDED\" field.",
                    "type": "string"
                },
                "HOURS": {
                    "description": "HOURS holds the value of the \"HOURS\" field.",
                    "type": "integer"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the ActivityQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.ActivityEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                }
            }
        },
        "ent.ActivityEdges": {
            "type": "object",
            "properties": {
                "actiAgen": {
                    "description": "ActiAgen holds the value of the acti_agen edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Agency"
                },
                "actiPlace": {
                    "description": "ActiPlace holds the value of the acti_place edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Place"
                },
                "actiStud": {
                    "description": "ActiStud holds the value of the acti_stud edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Student"
                },
                "actiYear": {
                    "description": "ActiYear holds the value of the acti_year edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Year"
                }
            }
        },
        "ent.Agency": {
            "type": "object",
            "properties": {
                "AGENCY": {
                    "description": "AGENCY holds the value of the \"AGENCY\" field.",
                    "type": "string"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the AgencyQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.AgencyEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                }
            }
        },
...
        "ent.Continent": {
            "type": "object",
            "properties": {
                "continent": {
                    "description": "Continent holds the value of the \"continent\" field.",
                    "type": "string"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the ContinentQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.ContinentEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                }
            }
        },
...
        "ent.Country": {
            "type": "object",
            "properties": {
                "country": {
                    "description": "Country holds the value of the \"country\" field.",
                    "type": "string"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the CountryQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.CountryEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                }
            }
        },
...
        "ent.Course": {
            "type": "object",
            "properties": {
                "course": {
                    "description": "Course holds the value of the \"course\" field.",
                    "type": "string"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the CourseQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.CourseEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                }
            }
        },
        "ent.CourseEdges": {
            "type": "object",
            "properties": {
                "courDegr": {
                    "description": "CourDegr holds the value of the cour_degr edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Degree"
                },
                "courFacu": {
                    "description": "CourFacu holds the value of the cour_facu edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Faculty"
                },
                "courInst": {
                    "description": "CourInst holds the value of the cour_inst edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Institution"
                }
            }
        },
        "ent.Degree": {
            "type": "object",
            "properties": {
                "degree": {
                    "description": "Degree holds the value of the \"degree\" field.",
                    "type": "string"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the DegreeQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.DegreeEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                }
            }
        },
        "ent.DegreeEdges": {
            "type": "object",
            "properties": {
                "degrCour": {
                    "description": "DegrCour holds the value of the degr_cour edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Course"
                    }
                },
                "degrStud": {
                    "description": "DegrStud holds the value of the degr_stud edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Student"
                    }
                }
            }
        },
        "ent.Faculty": {
            "type": "object",
            "properties": {
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the FacultyQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.FacultyEdges"
                },
                "faculty": {
                    "description": "Faculty holds the value of the \"faculty\" field.",
                    "type": "string"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                }
            }
        },
        "ent.FacultyEdges": {
            "type": "object",
            "properties": {
                "facuCour": {
                    "description": "FacuCour holds the value of the facu_cour edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Course"
                    }
                },
                "facuProf": {
                    "description": "FacuProf holds the value of the facu_prof edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Professor"
                    }
                }
            }
        },
        "ent.Gender": {
            "type": "object",
            "properties": {
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the GenderQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.GenderEdges"
                },
                "gender": {
                    "description": "Gender holds the value of the \"gender\" field.",
                    "type": "string"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                }
            }
        },
        "ent.GenderEdges": {
            "type": "object",
            "properties": {
                "gendStud": {
                    "description": "GendStud holds the value of the gend_stud edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Student"
                    }
                }
            }
        },
        "ent.Institution": {
            "type": "object",
            "properties": {
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the InstitutionQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.InstitutionEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                },
                "institution": {
                    "description": "Institution holds the value of the \"institution\" field.",
                    "type": "string"
                }
            }
        },
        "ent.InstitutionEdges": {
            "type": "object",
            "properties": {
                "instCour": {
                    "description": "InstCour holds the value of the inst_cour edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Course"
                    }
                }
            }
        },
        "ent.Place": {
            "type": "object",
            "properties": {
                "PLACE": {
                    "description": "PLACE holds the value of the \"PLACE\" field.",
                    "type": "string"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the PlaceQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.PlaceEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                }
            }
        },
...
        "ent.Prefix": {
            "type": "object",
            "properties": {
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the PrefixQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.PrefixEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                },
                "prefix": {
                    "description": "Prefix holds the value of the \"prefix\" field.",
                    "type": "string"
                }
            }
        },
        "ent.PrefixEdges": {
            "type": "object",
            "properties": {
                "preProf": {
                    "description": "PreProf holds the value of the pre_prof edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Professor"
                    }
                },
                "prefStud": {
                    "description": "PrefStud holds the value of the pref_stud edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Student"
                    }
                }
            }
        },
        "ent.Professor": {
            "type": "object",
            "properties": {
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the ProfessorQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.ProfessorEdges"
                },
                "email": {
                    "description": "Email holds the value of the \"email\" field.",
                    "type": "string"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                },
                "name": {
                    "description": "Name holds the value of the \"name\" field.",
                    "type": "string"
                },
                "tel": {
                    "description": "Tel holds the value of the \"tel\" field.",
                    "type": "string"
                }
            }
        },
        "ent.ProfessorEdges": {
            "type": "object",
            "properties": {
                "profFacu": {
                    "description": "ProfFacu holds the value of the prof_facu edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Faculty"
                },
                "profPre": {
                    "description": "ProfPre holds the value of the prof_pre edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Prefix"
                },
                "profPros": {
                    "description": "ProfPros holds the value of the prof_pros edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Professorship"
                }
            }
        },
        "ent.Professorship": {
            "type": "object",
            "properties": {
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the ProfessorshipQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.ProfessorshipEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                },
                "professorship": {
                    "description": "Professorship holds the value of the \"professorship\" field.",
                    "type": "string"
                }
            }
        },
        "ent.ProfessorshipEdges": {
            "type": "object",
            "properties": {
                "prosProf": {
                    "description": "ProsProf holds the value of the pros_prof edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Professor"
                    }
                }
            }
        },
        "ent.Province": {
            "type": "object",
            "properties": {
                "district": {
                    "description": "District holds the value of the \"district\" field.",
                    "type": "string"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the ProvinceQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.ProvinceEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                },
                "postal": {
                    "description": "Postal holds the value of the \"postal\" field.",
                    "type": "integer"
                },
                "province": {
                    "description": "Province holds the value of the \"province\" field.",
                    "type": "string"
                },
                "subdistrict": {
                    "description": "Subdistrict holds the value of the \"subdistrict\" field.",
                    "type": "string"
                }
            }
        },
        "ent.ProvinceEdges": {
            "type": "object",
            "properties": {
                "distStud": {
                    "description": "DistStud holds the value of the dist_stud edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Student"
                    }
                },
                "postStud": {
                    "description": "PostStud holds the value of the post_stud edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Student"
                    }
                },
                "provCont": {
                    "description": "ProvCont holds the value of the prov_cont edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Continent"
                },
                "provCoun": {
                    "description": "ProvCoun holds the value of the prov_coun edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Country"
                },
                "provRegi": {
                    "description": "ProvRegi holds the value of the prov_regi edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Region"
                },
                "provStud": {
                    "description": "ProvStud holds the value of the prov_stud edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Student"
                    }
                },
                "subdStud": {
                    "description": "SubdStud holds the value of the subd_stud edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Student"
                    }
                }
            }
        },
        "ent.Region": {
            "type": "object",
            "properties": {
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the RegionQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.RegionEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                },
                "name": {
                    "description": "Name holds the value of the \"name\" field.",
                    "type": "string"
                }
            }
        },
...
        "ent.Results": {
            "type": "object",
            "properties": {
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the ResultsQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.ResultsEdges"
                },
                "grade": {
                    "description": "Grade holds the value of the \"grade\" field.",
                    "type": "number"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                }
            }
        },
        "ent.ResultsEdges": {
            "type": "object",
            "properties": {
                "resuStud": {
                    "description": "ResuStud holds the value of the resu_stud edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Student"
                },
                "resuSubj": {
                    "description": "ResuSubj holds the value of the resu_subj edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Subject"
                },
                "resuTerm": {
                    "description": "ResuTerm holds the value of the resu_term edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Term"
                },
                "resuYear": {
                    "description": "ResuYear holds the value of the resu_year edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Year"
                }
            }
        },
        "ent.Student": {
            "type": "object",
            "properties": {
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the StudentQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.StudentEdges"
                },
                "email": {
                    "description": "Email holds the value of the \"email\" field.",
                    "type": "string"
                },
                "fname": {
                    "description": "Fname holds the value of the \"fname\" field.",
                    "type": "string"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                },
                "lname": {
                    "description": "Lname holds the value of the \"lname\" field.",
                    "type": "string"
                },
                "recent_address": {
                    "description": "RecentAddress holds the value of the \"recent_address\" field.",
                    "type": "string"
                },
                "schoolname": {
                    "description": "Schoolname holds the value of the \"schoolname\" field.",
                    "type": "string"
                },
                "telephone": {
                    "description": "Telephone holds the value of the \"telephone\" field.",
                    "type": "integer"
                }
            }
        },
        "ent.StudentEdges": {
            "type": "object",
            "properties": {
                "studActi": {
                    "description": "StudActi holds the value of the stud_acti edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Activity"
                    }
                },
                "studDegr": {
                    "description": "StudDegr holds the value of the stud_degr edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Degree"
                },
                "studDist": {
                    "description": "StudDist holds the value of the stud_dist edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Province"
                },
                "studGend": {
                    "description": "StudGend holds the value of the stud_gend edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Gender"
                },
                "studPost": {
                    "description": "StudPost holds the value of the stud_post edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Province"
                },
                "studPref": {
                    "description": "StudPref holds the value of the stud_pref edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Prefix"
                },
                "studProv": {
                    "description": "StudProv holds the value of the stud_prov edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Province"
                },
                "studResu": {
                    "description": "StudResu holds the value of the stud_resu edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Results"
                    }
                },
                "studSubd": {
                    "description": "StudSubd holds the value of the stud_subd edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Province"
                }
            }
        },
        "ent.Subject": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "Code holds the value of the \"code\" field.",
                    "type": "integer"
                },
                "creditpiont": {
                    "description": "Creditpiont holds the value of the \"creditpiont\" field.",
                    "type": "integer"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the SubjectQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.SubjectEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                },
                "subjects": {
                    "description": "Subjects holds the value of the \"subjects\" field.",
                    "type": "string"
                }
            }
        },
...
        "ent.Year": {
            "type": "object",
            "properties": {
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the YearQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.YearEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                },
                "years": {
                    "description": "Years holds the value of the \"years\" field.",
                    "type": "integer"
                }
            }
        },
...
        "gin.H": {
            "type": "object",
            "additionalProperties": true
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "BasicAuth": {
            "type": "basic"
        },
        "OAuth2AccessCode": {
            "type": "oauth2",
            "flow": "accessCode",
            "authorizationUrl": "https://example.com/oauth/authorize",
            "tokenUrl": "https://example.com/oauth/token",
            "scopes": {
                "admin": " Grants read and write access to administrative information"
            }
        },
        "OAuth2Application": {
            "type": "oauth2",
            "flow": "application",
            "tokenUrl": "https://example.com/oauth/token",
            "scopes": {
                "admin": " Grants read and write access to administrative information",
                "write": " Grants write access"
            }
        },
        "OAuth2Implicit": {
            "type": "oauth2",
            "flow": "implicit",
            "authorizationUrl": "https://example.com/oauth/authorize",
            "scopes": {
                "admin": " Grants read and write access to administrative information",
                "write": " Grants write access"
            }
        },
        "OAuth2Password": {
            "type": "oauth2",
            "flow": "password",
            "tokenUrl": "https://example.com/oauth/token",
            "scopes": {
...
}